Déplacement
juste déplacement des fichiers pour une arborescence plus logique

+++ sources/item/ItemSet.cpp
@@ -2,10 +2,10 @@
 
 ItemSet::ItemSet(std::vector<std::vector<Item*>>& items, std::vector<std::vector<int>>& quantity)
 {
-    for(unsigned int i = 0; i < items.size(); i++)
+    for(size_t i = 0; i < items.size(); i++)
     {
         std::vector<ItemQuantity*> vec;
-        for(unsigned int j = 0; j < items[i].size(); j++)
+        for(size_t j = 0; j < items[i].size(); j++)
         {
             ItemQuantity* n = new ItemQuantity;
             n->item = items[i][j];
@@ -18,9 +18,9 @@ ItemSet::ItemSet(std::vector<std::vector<Item*>>& items, std::vector<std::vector
 
 ItemSet::~ItemSet()
 {
-    for (unsigned int i = 0; i < m_set.size(); i++)
+    for (size_t i = 0; i < m_set.size(); i++)
     {
-        for(unsigned int j = 0; j < m_set[i].size(); j++)
+        for(size_t j = 0; j < m_set[i].size(); j++)
         {
             delete m_set[i][j];
         }
@@ -40,10 +40,10 @@ std::vector<std::vector<ItemQuantity*>> ItemSet::getCopySet()
 std::ostream& operator<<(std::ostream& out, ItemSet& s)
 {
 	out << "[";
-	for (unsigned int i = 0; i < s.m_set.size(); i++)
+	for (size_t i = 0; i < s.m_set.size(); i++)
     {
         if(s.m_set[i].size() > 1) out << "(";
-        for(unsigned int j = 0; j < s.m_set[i].size(); j++)
+        for(size_t j = 0; j < s.m_set[i].size(); j++)
         {
             out << "(\"" << s.m_set[i][j]->item->getName() << "\"," << s.m_set[i][j]->quantity << ")";
         }
